requirements.txt
# local package
-e .

# external requirements
appnope==0.1.2
backcall==0.2.0
cycler==0.10.0
decorator==4.4.2
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.11
jupyter-core==4.7.0
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.5
pandas==1.2.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prompt-toolkit==3.0.10
ptyprocess==0.7.0
py4j==0.10.9
Pygments==2.7.3
pyparsing==2.4.7
pyspark==3.0.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
pyzmq==20.0.0
six==1.15.0
SQLAlchemy==1.3.22
tornado==6.1
traitlets==5.0.5
wcwidth==0.2.5
